requirements.txt
flask
werkzeug
gtts
groq
python-dotenv
gunicorn
